chatbot/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
sqlalchemy==1.4.56
alembic==1.11.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
langchain-groq==0.1.0
pydantic==1.10.10
databases==0.7.3